Airflow/M3S1-Airflow_pipeline/requirements.txt
apache-airflow==2.10.0
apache-airflow-providers-postgres==5.12.0
loguru==0.7.2
numpy==2.1.0
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy>=1.4.36
psycopg2-binary==2.9.9